requirements.txt
scikit-learn
python-Levenshtein
tqdm
google-generativeai
